User controller test

+++ src/main/java/ZCW/ChatApp/services/UserService.java
@@ -91,12 +91,17 @@ public class UserService {
         return userRepo.save(original);
     }
 
+    // Update username, password for user
+
 
     // DELETE
     //=============================================================================
     public Boolean deleteUser(Long id){
-        userRepo.deleteById(id);
-        return true;
+        if(findById(id).isPresent()){
+            userRepo.deleteById(id);
+            return true;
+        }
+        else return false;
     }
 
     public Boolean deleteAll(){
